requirements-capitalai.txt
# requirements-capitalai.txt
# CapitalAI additions on top of the existing Crawl4AI fork dependencies.
# Install AFTER: pip install -e ".[all]" && crawl4ai-setup
#
# Usage:
#   pip install -r requirements-capitalai.txt

# CLI + output
httpx>=0.27
rich>=13.7
typer>=0.12
python-dotenv>=1.0

# PDF report (pure Python, zero system deps, works on Windows/Mac/Linux)
# Do NOT use WeasyPrint — requires GTK3/Cairo system libs, breaks on Windows
reportlab>=4.2.0

# HTML email writer — no extra deps (stdlib only: pathlib, datetime)
# crawl4ai already provides: playwright, beautifulsoup4, lxml, requests
